requirements.txt
fastapi
uvicorn[standard]
pydicom
oss2
python-dotenv
PyYAML
python-multipart
